+++ __init__.py
@@ -12,6 +12,7 @@ import bpy
 from bpy.props import PointerProperty
 
 from .operators import (
+    PCG_OT_add_selected_colliders,
     PCG_OT_create_cable_from_object_chain,
     PCG_OT_create_cable_from_selection,
     PCG_OT_create_cables_from_out_mid_in,
@@ -32,6 +33,7 @@ _CLASSES = (
     PCG_OT_create_cables_from_out_mid_in,
     PCG_OT_make_cable_dynamic,
     PCG_OT_remove_cable_dynamics,
+    PCG_OT_add_selected_colliders,
     PCG_PT_cable_panel,
 )